requirements.txt
# Core emotion detection
transformers>=4.35.0
torch>=2.0.0
vaderSentiment>=3.3.2

# TTS engines
gTTS>=2.3.2
pyttsx3>=2.90
pydub>=0.25.1

# Web interface
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Utilities
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
requests>=2.31.0
python-dotenv>=1.0.0
